+++ Modules/Intent.Modules.Application.ServiceImplementations.CRUD/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 5.0.19
 
 - Improvement: Added support for ProjectTo query implementations for DBContext implementation CRUD Scenarios.
+- Improvement: Comparison expressions with `bool` will generate a simplifed expression (e.g. `x.IsActive` instead `x.IsActive == true`)
 
 ### Version 5.0.18