+++ Modules/Intent.Modules.MediatR.DomainEvents/release-notes.md
@@ -2,6 +2,7 @@
 
 - Improvement: Improved code quality by making `DomainEventService.GetNotificationCorrespondingToDomainEvent` static.
 - Improvement: Added support for ProjectTo query implementations for DBContext implementation CRUD Scenarios.
+- Improvement: Comparison expressions with `bool` will generate a simplifed expression (e.g. `x.IsActive` instead `x.IsActive == true`)
 
 ### Version 5.0.18